Update README to describe tegrarcm readbct command
Recent versions of tegrarcm now have the capability to read the BCT from the
device without requiring any HW-specific binaries, nor code already installed
on the device. Mention this in the README, since it discusses ways to extract
a BCT from a device.

Also, fix a typo in the Linux MMC block device name that contains the BCT.

Signed-off-by: Stephen Warren <swarren@nvidia.com>