post.mk: Create dependency makefile by makefile rule
Creating the eMMC image requires creating the BCT file and the
bootloader image first.  These dependencies are written to a makefile
snippet that is created by the gen-image-deps.sh script. But that
script is only invoked when the rule that creates the eMMC image is run,
which is too late for make to take account of the freshly generated
dependencies in the current invocation.

Add a rule to post.mk that tells make how to create the dependencies
makefile; with that in place, make automatically makes the dependency
makefile before executing any other rules.

Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
(swarren, minor commit description rewording)
Signed-off-by: Stephen Warren <swarren@nvidia.com>